Refactor IUserController and UserController into package Controller.UserController
