+++ src/main/java/Controller/UserController/IUserController.java
@@ -1,4 +1,4 @@
-package Controller;
+package Controller.UserController;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
